+++ CHANGELOG.md
@@ -9,11 +9,17 @@ and to the [CHANGELOG recommendations](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- Add currency on Booking amount error message and change min listing price to 1
+- Add fees help in sonata admin for bank wire
+
 ### Fixed
 - Fix bug "Duplicate error message on new booking"    
+- Fix bug : "Bookings refusing while booking acceptation"
 
 ### Changed
 - Update documentation
+
 
 ## [0.1.0] - 2016-03-23